Moved files around and switched to webpack.
