+++ package.json
@@ -1,0 +1,23 @@
+{
+  "private": true,
+  "sideEffects": false,
+  "devDependencies": {
+    "@types/chart.js": "^2.9.30",
+    "autoprefixer": "^10.2.4",
+    "chart.js": "^2.9.4",
+    "postcss": "^8.2.5",
+    "postcss-cli": "^8.3.1",
+    "tailwindcss": "^2.0.3",
+    "typescript": "^4.1.3",
+    "webpack": "^5.21.2",
+    "webpack-cli": "^4.5.0"
+  },
+  "scripts": {
+    "build": "npm run build:css && npm run build:html && npm run build:js",
+    "build:css": "postcss css/style.css -o dist/style.css",
+    "build:html": "npm run build:html_details",
+    "build:html_details": "python render_template.py details.jinja2 dist/details.html",
+    "build:js": "npm run build:ts && webpack",
+    "build:ts": "tsc"
+  }
+}
