+++ src/dashboard/analysis.ts
