+++ templates/details.jinja2
@@ -43,6 +43,6 @@
     <canvas id="growth-factor"></canvas>
 </div>
 
-<script src="report.js"></script>
+<script src="details.js"></script>
 
 {% endblock %}
