--- web/package.json
@@ -1,20 +1,0 @@
-{
-  "devDependencies": {
-    "@types/chart.js": "^2.9.30",
-    "autoprefixer": "^10.2.3",
-    "browserify": "^17.0.0",
-    "chart.js": "^2.9.4",
-    "del": "^6.0.0",
-    "gulp": "^4.0.2",
-    "gulp-postcss": "^9.0.0",
-    "gulp-sourcemaps": "^3.0.0",
-    "gulp-terser": "^2.0.1",
-    "postcss": "^8.2.4",
-    "tailwindcss": "^2.0.2",
-    "tsify": "^5.0.2",
-    "typescript": "^4.1.3",
-    "vinyl-buffer": "^1.0.1",
-    "vinyl-paths": "^3.0.1",
-    "vinyl-source-stream": "^2.0.0"
-  }
-}
