+++ webpack.config.js
@@ -1,0 +1,18 @@
+const path = require('path');
+
+module.exports = {
+    mode: 'development',
+    optimization: {
+        usedExports: true,
+    },
+    externals: {
+        'chart.js': 'Chart',
+    },
+    entry: {
+        details: './build/lib/dashboard/pages/details.js'
+    },
+    output: {
+        filename: '[name].js',
+        path: path.resolve(__dirname, 'dist')
+    }
+};
